+++ index.js
@@ -1,6 +1,4 @@
 /*
- * byte-data
- * Pack and unpack binary data.
  * Copyright (c) 2017-2018 Rafael da Silva Rocha.
  * https://github.com/rochars/byte-data
  *
@@ -16,11 +14,12 @@ const packer = require("./lib/packer");
  * Write a number or fixed-length string to a byte buffer.
  * @param {number|string} value The value.
  * @param {!Object} theType The type definition.
+ * @param {!boolean=} allowOverflow True to truncate values on overflow/underflow.
  * @return {!Array<number|string>}
  * @throws {Error} If the type definition is not valid.
  */
-function pack(value, theType) {
-    packer.setUp(theType);
+function pack(value, theType, allowOverflow=false) {
+    packer.setUp(theType, allowOverflow);
     let packed = [];
     if (value === undefined) {
         return packed;
@@ -33,11 +32,12 @@ function pack(value, theType) {
  * Read a number or a fixed-length string from a byte buffer.
  * @param {!Array<number|string>|!Uint8Array} buffer An array of bytes.
  * @param {!Object} theType The type definition.
+ * @param {!boolean=} allowOverflow True to truncate values on overflow/underflow.
  * @return {number|string|null}
  * @throws {Error} If the type definition is not valid.
  */
-function unpack(buffer, theType) {
-    packer.setUp(theType);
+function unpack(buffer, theType, allowOverflow=false) {
+    packer.setUp(theType, allowOverflow);
     let values = packer.fromBytes(
         buffer.slice(0, theType["offset"]), theType);
     return values ? values[0] : theType["char"] ? "" : null;
@@ -47,11 +47,12 @@ function unpack(buffer, theType) {
  * Write an array of numbers or a string to a byte buffer.
  * @param {!Array<number|string>} values The values.
  * @param {!Object} theType The type definition.
+ * @param {!boolean=} allowOverflow True to truncate values on overflow/underflow.
  * @return {!Array<number|string>}
  * @throws {Error} If the type definition is not valid.
  */
-function packArray(values, theType) {
-    packer.setUp(theType);
+function packArray(values, theType, allowOverflow=false) {
+    packer.setUp(theType, allowOverflow);
     // Fix strings with bad length in the array
     if (theType["char"]) {
         let len = values.length;
@@ -66,11 +67,12 @@ function packArray(values, theType) {
  * Read an array of numbers or a string from a byte buffer.
  * @param {!Array<number|string>|!Uint8Array} buffer The byte array.
  * @param {!Object} theType The type definition.
+ * @param {!boolean=} allowOverflow True to truncate values on overflow/underflow.
  * @return {!Array<number|string>|number}
  * @throws {Error} If the type definition is not valid.
  */
-function unpackArray(buffer, theType) {
-    packer.setUp(theType);
+function unpackArray(buffer, theType, allowOverflow=false) {
+    packer.setUp(theType, allowOverflow);
     return packer.fromBytes(buffer, theType);
 }