+++ test/src/pack/bool.js
@@ -12,17 +12,17 @@ describe('booleans to bytes', function() {
     // booleans
     it('should turn 1 value to 1 booolean bin (6)', function() {
         assert.deepEqual(
-            byteData.packArray([6], byteData.types.bool),
+            byteData.packArray([6], byteData.types.bool, true),
             [1]);
     });
     it('should turn 1 value to 1 booolean hex (6)', function() {
         assert.deepEqual(
-            byteData.packArray([6], byteData.types.bool),
+            byteData.packArray([6], byteData.types.bool, true),
             [1]);
     });
     it('should turn 1 value to 1 booolean decimal (6)', function() {
         assert.deepEqual(
-            byteData.packArray([6], byteData.types.bool),
+            byteData.packArray([6], byteData.types.bool, true),
             [1]);
     });
     it('should turn 1 value to 1 booolean bin (0)', function() {