+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v9.0.0 (unreleased)
 - dist included in npm package.
 - change: bytes only in base 10 for input and output
+- change: types in byteData.types
+- change: throw errors on overflow and underflow
 
 ## v8.0.3 (2018-06-11)
 - fix: webpack.config so no dependency dist is used in the bundle.